# openapi/v1.yml
openapi: '3.1.0'
info:
  title: TwextHub API
  version: '1.0.0'
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: >
    TwextHub is a lightweight registry for publishing and discovering
    Twext-compiled TurboWarp extensions ("Twexts"). Packages are namespaced
    to their publishing account (`@namespace/id`) and versioned with SemVer.
    Publishing sends a gzipped tarball of the Twext project (`twext.yml` plus
    sources); the hub validates the manifest and compiles the extension
    itself in a sandboxed build, so the reviewed source is exactly what
    produces the served blob.

    The `/v1` surface documented here is stable as of 1.0.0. Breaking changes
    require a new major version; additive changes bump the minor (see
    CHANGELOG.md).

    Every authenticated (`Userspace`/`Adminspace`) endpoint requires the
    caller to have accepted the current Terms of Service; a caller who
    hasn't gets a `403` (see `Forbidden`) until they patch their own user with
    `termsAcceptedVersion` set to the `version` that `GET /terms` reports. A
    stale version is a `422`, so an acceptance is always recorded against text
    the account was shown. Exemptions that stay available without
    re-acceptance: `DELETE /sessions/current`, `GET /me`, the acceptance
    itself, password-only user updates, and account deletion.

    Bearer tokens come in two kinds. **Sessions** are created by
...
    every scope. **Automation tokens** are created explicitly via
...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /admin/terms:
    patch:
      summary: Publish a new Terms of Service version
      operationId: updateTerms
      tags: [Adminspace]
      description: Replaces the current Terms of Service body and bumps `version`. Previously-accepted versions no longer satisfy the acceptance gate.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, description: Markdown source. }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LegalDocument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /admin/privacy:
    patch:
      summary: Publish a new Privacy Policy version
      operationId: updatePrivacyPolicy
      tags: [Adminspace]
      description: Replaces the current Privacy Policy body and bumps `version`. Does not itself require re-acceptance (only Terms updates gate `Userspace`/`Adminspace` calls).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, description: Markdown source. }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LegalDocument' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /admin/audit:
    get:
      summary: Append-only audit trail of privileged actions
      operationId: listAuditLog
      tags: [Adminspace]
      description: >
        Records publishes, review decisions, yanks, deprecations, tag and
        owner changes, quota changes, role changes, and access grants,
        newest first.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Dir'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated audit entries
          content:
            application/json:
              schema:
                type: object
                required: [data, _links]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEntry' }
                  _links: { $ref: '#/components/schemas/PageLinks' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/users/{namespace}/quota:
    parameters:
      - $ref: '#/components/parameters/NamespacePath'
    get:
      summary: Read an account's storage usage and quota
      operationId: getUserQuota
      tags: [Adminspace]
      responses:
        '200':
          description: Current usage
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Quota' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Set an account's storage quota override
      operationId: setUserQuota
      tags: [Adminspace]
      description: '`null` resets the account to the instance default.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [maxBlobBytes]
              properties:
                maxBlobBytes: { type: ['integer', 'null'], minimum: 1 }
      responses:
        '200':
          description: Updated quota
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Quota' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
  /admin/config:
    get:
      summary: Read the settings an admin may change
      operationId: getAdminConfig
      tags: [Adminspace]
      description: >
        The instance's own settings, limited to the ones an admin is allowed to
        change from the interface. The database URL, the API root, the port and
        the blob store location are not among them: a form is an easier way to
        stop an instance answering than a hand-edited file, and the API root in
        particular would 404 the route the admin is using to fix it.

        `editable` is false when the configuration file cannot hold a change.
        In a container that means the file sits in the instance's own writable
        layer rather than on a volume, so a change would be discarded when the
        instance is recreated. A host install writing to its own filesystem is
        left alone: a recreate does not discard that. `reason` explains the
        state, and `PUT` refuses with 409 until an operator mounts the file.
        Reported either way, so the interface can explain rather than showing a
        form that quietly does nothing.
      responses:
        '200':
          description: Current settings
          content:
            application/json:
              schema:
                type: object
                required: [editable, configPath, settings]
                properties:
                  editable:
                    type: boolean
                    description: Whether a change written here would survive a restart.
                  reason:
                    type: ['string', 'null']
                    description: Why the file is read-only, when it is.
                  configPath:
                    type: string
                  settings:
                    type: array
                    items: { $ref: '#/components/schemas/ServerSetting' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      summary: Change the instance's settings
      operationId: updateAdminConfig
      tags: [Adminspace]
      description: >
        Applies the submitted settings to the configuration file and to the
        running instance, so a setting the process reads per request takes
        effect on the next request. `restartRequired` lists any changed setting
        that is only read at startup.

        The file is edited as a YAML document, so comments and key order the
        operator wrote survive. A setting already at the submitted value is not
        written and does not appear in `changed`. A key outside the offered set
        is refused by name rather than ignored, so a caller is never left
        believing it changed something it did not.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [settings]
              properties:
                settings:
                  type: object
                  additionalProperties: true
                  example:
                    limits.maxProfileImageBytes: 4194304
                    pagination.maxLimit: 100
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  changed:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        before: {}
                        after: {}
                  restartRequired:
                    type: array
                    items: { type: string }
                  settings:
                    type: array
                    items: { $ref: '#/components/schemas/ServerSetting' }
        '409':
          description: The configuration file cannot hold a change
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/notifications:
    post:
      summary: Broadcast a notification to every account
      operationId: broadcastNotification
      tags: [Adminspace]
      description: >
        Fan-out at insert time: every account (past and present at the moment
        of the call) gets its own `broadcast` notification row. Use for
        registry-wide notices like maintenance windows or policy changes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
                  minLength: 1
                  maxLength: 280
                  pattern: '\S'
      responses:
        '201':
          description: Delivered
          content:
            application/json:
              schema:
                type: object
                required: [created]
                properties:
                  created:
                    { type: integer, description: Number of accounts the message was written to. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
...
components:
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: >
        Opaque position from a previous page, taken from that page's
        `_links.next` or `_links.prev` rather than assembled by hand.
    Dir:
      name: dir
      in: query
      schema: { type: string, enum: [back] }
      description: >
        Set to `back` to page towards the rows before the cursor. The server only
        looks for a cursor from a list sorted the other way round; anything else
        is a `400`. A client never needs it, because every page's `_links.prev`
        already sets it.
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, default: 20 }
      description: Values above the server cap are clamped to the cap, not rejected. The default cap is 50.
    NamespacePath:
      name: namespace
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Namespace' }
...
  schemas:
    Namespace:
      type: string
      pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
...
    LegalDocument:
      type: object
      required: [version, body, updatedAt]
      properties:
        version: { type: integer, description: Increments on every publish of a new version. }
        body: { type: string, description: Markdown source. }
        updatedAt: { type: string, format: date-time }
...
    AuditEntry:
      type: object
      required: [id, actor, action, target, detail, createdAt]
      properties:
        id: { type: string }
        actor: { type: string, description: 'Actor namespace, or `system`.' }
        action:
          type: string
          description: >-
            e.g. version.publish, version.approve, version.reject,
            version.yank, version.deprecate, tag.set, tag.remove, owner.invite,
            owner.accept, owner.withdraw, owner.remove, access.grant,
            access.revoke, quota.set, extension.transfer.request,
            extension.transfer.withdraw, extension.transfer,
            role.change, extension.delete, org.create, org.update, org.delete,
            organization.owner.add, organization.owner.remove,
            organization.webhook.create, organization.webhook.delete
        target:
          type: object
          properties:
            namespace: { type: ['string', 'null'] }
            id: { type: ['string', 'null'] }
            version: { type: ['string', 'null'] }
        detail: { type: object }
        createdAt: { type: string, format: date-time }
    Quota:
      type: object
      required: [namespace, blobBytes, maxBlobBytes]
      properties:
        namespace: { $ref: '#/components/schemas/Namespace' }
        blobBytes:
          {
            type: integer,
            description: Bytes currently charged to the account (blobs plus retained sources).,
          }
        maxBlobBytes:
          {
            type: ['integer', 'null'],
            description: Per-account override; `null` means the instance default.,
          }
    PageLinks:
      type: object
      description: >
        The pages either side of this one, as complete URLs. Each already carries
        the cursor, the limit and the filters it was built from, so a client
        follows a link rather than assembling a query; a null means that
        direction is the end of the list.
      required: [self, next, prev]
      properties:
        self:
          type: string
          description: The request as it was made.
        next: { type: ['string', 'null'] }
        prev: { type: ['string', 'null'] }
    ServerSetting:
      type: object
      required: [key, label, type, restartRequired, value]
      properties:
        key:
          type: string
          description: Dotted path in the configuration file, e.g. `limits.maxBlobBytes`.
        label: { type: string }
        help: { type: ['string', 'null'] }
        type:
          type: string
          enum: [string, number, bytes, boolean, url, origins]
        min: { type: ['integer', 'null'] }
        max: { type: ['integer', 'null'] }
        restartRequired:
          type: boolean
          description: >
            True when the running process only reads this at startup, so the
            change takes effect after a restart rather than on the next request.
        value: {}
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: 'about:blank' }
        title: { type: string, example: 'Validation Error' }
        status: { type: integer, example: 422 }
        detail: { type: string }
        instance: { type: string, format: uri }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Forbidden:
      description: >
        Authenticated, but not allowed to perform this action — e.g. not the
        resource owner or an admin, a namespace mismatch on publish, the
        current Terms of Service haven't been accepted yet
        (see `termsAcceptedVersion` on `PATCH /users/{namespace}`), or a
        credential missing the scope the route needs. The `detail` names the
        scope. An automation token on `/admin/*` or a version review also needs
        the `admin` scope on top of the account being an admin.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    NotFound:
      description: Resource not found
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    UnprocessableEntity:
      description: Semantically invalid input (e.g. non-SemVer version, id mismatch)
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer